merge floats

+++ libft/ft_itoa.c
@@ -3,16 +3,16 @@
 /*                                                        :::      ::::::::   */
 /*   ft_itoa.c                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: coclayto <coclayto@42.fr>                  +#+  +:+       +#+        */
+/*   By: coclayto <coclayto@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/04/30 20:33:26 by coclayto          #+#    #+#             */
-/*   Updated: 2019/04/30 20:33:29 by coclayto         ###   ########.fr       */
+/*   Updated: 2020/03/01 10:50:57 by coclayto         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "includes/libft.h"
 
-static size_t	str_len(int n)
+static size_t	str_len(long long n)
 {
 	size_t		i;
 
@@ -22,7 +22,7 @@ static size_t	str_len(int n)
 	return (i);
 }
 
-char			*ft_itoa(int n)
+char			*ft_itoa(long long n)
 {
 	char			*str;
 	size_t			len;
